requirements.txt
paho-mqtt
docker
geocoder
ipaddress
pytest
pytest-mock
pytest-cov